Revert "Simplify forms and controller code"
This reverts commit 35f341fd4008de2e295f434039bc03f72b66e61a.

+++ webapp/src/main/java/ar/edu/itba/paw/webapp/form/AddCategoryForm.java
@@ -7,17 +7,17 @@ import javax.validation.constraints.Size;
 
 public class AddCategoryForm {
     @NotNull
-    private Long restaurantId;
+    private Integer restaurantId;
 
     @NotBlank
     @Size(max = 50)
     private String name;
 
-    public Long getRestaurantId() {
+    public Integer getRestaurantId() {
         return restaurantId;
     }
 
-    public void setRestaurantId(long restaurantId) {
+    public void setRestaurantId(int restaurantId) {
         this.restaurantId = restaurantId;
     }
