+++ webapp/src/main/java/ar/edu/itba/paw/webapp/form/CartItem.java
@@ -8,7 +8,7 @@ import javax.validation.constraints.Size;
 public class CartItem {
 
     @NotNull
-    private Long productId;
+    private Integer productId;
 
     @Min(1)
     @Max(100)
@@ -18,11 +18,11 @@ public class CartItem {
     @Size(max = 120)
     private String comment;
 
-    public long getProductId() {
+    public int getProductId() {
         return productId;
     }
 
-    public void setProductId(long productId) {
+    public void setProductId(int productId) {
         this.productId = productId;
     }
